Implemented mechanics of shield durability
- While blocking, your shield takes the damage instead of your health
- As the shield gets damaged, its overall damage reduction gets reduced and the player starts taking some of the damage too
- The shield will start regenerating after a delay and if the player is not blocking
- If the shield breaks, it won't start regenerating until after a longer delay
- The code (particularly in PlayerCombat) for this will likely get reorganized in the future because it's a bit messy right now

+++ Assets/Scripts/EnemyAttackHitbox.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class EnemyAttackHitbox : MonoBehaviour
 {
     public EnemyController enemyController;
-    [HideInInspector] public int AttackDamage;
+    [HideInInspector] public float AttackDamage;
     [HideInInspector] public bool HasHitPlayerYet;
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
